ntbtls-cli: Use union to access hostent addr.
* src/ntbtls-cli.c (connect_server): Use union.

--

Unfortunately, the API of gethostbyname was designed in old style of
C.  Modern compiler emits warnings for its access to h_addr of char *
on a machine of different alignment requirement for other
objects.  (It should be defined as void *, in new style.)

As a workaround, we use union here.

Signed-off-by: NIIBE Yutaka <gniibe@fsij.org>